Export `CD_LEVEL=<num>` instead of `PS1_PREFIX`; export `CD_DIRECTORY=<dir>`
Originally we used `PS1_PREFIX` with string value, which was directly
used in `PS1`. Now we define `CD_LEVEL` that contains how many nested
`TOOLSET cd` calls we've made. This seems to be more useful and it also
fallows similar pattern as `SHLVL`.

Introduced new environment variable `CD_DIRECTORY` which contains the
directory path of the original working directory used when `TOOLSET cd`
invoked the shell. This allows us to go back to this directory if we
change it. (`cd "${CD_DIRECTORY}"`)

These environment variables are defined when new shell or terminal
emulator is executed, but not when new Tmux window is created.