Making `colourOutput` field in config file optional
This behaviour makes much more sense since we are trying to respect
`NO_COLOR` environment variable. Reason for this is that the value set
in a configuration file would always override `NO_COLOR`, which is not
what we want. By having it optional we can override it if necessary.
This behaviour is consistent with <https://no-color.org>.