Reimplementing `COMMAND_WRAPPER_INVOKE_AS` environment variable
It's value overrides the name under which `command-wrapper` command was
executed, i.e. what is returned by `getProgName`.  See
`man/command-wrapper.1.md` section *ENVIRONMENT VARIABLES* for more
details.

Originally it was hardcoded in `getAppNames` function.  This worked
well, but in the future we want Command Wrapper to work as a library
where these things are configurable.  We already have such abstraction
in place for subcomman protocol.  This commit does the same for
environment variables that are used by the toolset executable itself.

Some documentation was added as part of this commit.