Generalised `menuTool` and `terminalEmulator` configuration for `cd`
Value of `QUERY` passed to `cd` via `--query=QUERY` is now passed to
`menuTool` (AKA finder tool) as an argument.  This allows us to do
stuff like:

```
TOOLSET_COMMAND cd --query=.config
```

In the above example the finder tool will start with `.config` as a
pre-filled query.  This is very useful when we want to pre-filter some
of the options.

Configuration field `terminalEmulator` is more general now as well.  It
accepts a command to start.  This way we can start a terminal emulator
with a different shell, the one that is also part of the `cd` subcommand
configuration, instead of default shell.

When starting new Tumux/Kitty window we now start the configured shell
as well. Also we managed to pass `CD_LEVEL` and `CD_DIRECTORY` to this
new shell/window.  This makes behaviour across modes consistent.

Dhall configs are consistent with internal `cd` types, therefore, there
is a need to bump import checksum.