Introduced completion option `--subcommand=SUBCOMMAND`
This is a first step to start supporting subcommand command line
completion.  At the moment only `--subcommand=help` gives any meaningful
results.  Any other value will return empty completion list.

In case of external subcommands plan is to support call them with
`--completion-config` which should give us a Dhall function which
constructs argument list which should be used to call completion on a
subcommand.  In other words the process for subcommand completion is:

1.  Call subcommand with `--completion-config`.
2.  Evaluate returned Dhall function int to a list of `ARGUMENTS`.
3.  Call subcommand with `ARGUMENTS`

We need to describe this in Subcommand Protocol as well.

Next steps:

*   Implement completion for rest of the internal subcommands.
*   Implement above protocol for
    `TOOLSET completion --subcommand=SUBCOMMAND`.
*   Make global `TOOLSET completion` aware of subcommands, and how to
    complete their arguments.