Parsing `--man` as a flag instead of option
Package `optparse-applicative` doesn't support options with optional
argument.  For that reason we had to implement `--man` as a flag, i.e.
option that doesn't take argument.

When argument is not specified `help --man` behaves in a similar way as
`help` on its own, i.e. it prints manual page for the current toolset.