Annotate alias description with aliased command
Purpose of this is to make it obvious what underlying command is being
called. For example, a lot of commands can be implemented using `exec`,
which has nice "transparency" property via `--print`, however that's
hard to use when one doesn't know that the command is alias for an
`exec` command.